Hotfix 0.1 + Small update on navbar
In this patch, I fixed some errors that were occurring in the footer and the page logo. I also racked my brain to create a new sidebar, which will include the contents with the old HTML and CSS pages, as well as the ones that will be created in the future. I ended up using JavaScript, tried some ideas in CSS, and in the end, I liked my own CSS creation better without using JavaScript.